feat: create a PlayerControlsScaffold
* Create a PlayerControlScaffold to define the playerOverlay
 layout
* Divide player components "MiddleControls", "BottomControls"
 into different files to make it more organized and easier to read
* Remove the customization modifier because it can be confusing and
each of your composable icons can have its own modifier set by whoever
creates it

Closes #32
